--- JoyboyCommunity/src/components/avatar/props.ts
@@ -1,4 +1,0 @@
-export type AvatarProps = {
-  userId: string;
-  source?: string;
-};
