+++ JoyboyCommunity/src/modules/feed/components/FloatingPostButton.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import styled from "styled-components/native";
 import { Feather } from "@expo/vector-icons";
 import { useNavigation } from "@react-navigation/native";
+import { RootStackNavigationProps } from "../../../types";
 
 const CircleButton = styled(Pressable)`
   width: 56px;
@@ -15,11 +16,12 @@ const CircleButton = styled(Pressable)`
 `;
 
 export default function FloatingPostButton() {
-  const navigation = useNavigation();
+  const navigation = useNavigation<RootStackNavigationProps>();
 
   const handleNavigation = () => {
     navigation.push("CreatePost");
   };
+
   return (
     <CircleButton onPress={handleNavigation}>
       <Feather name="plus" size={24} color="white" />
