+++ JoyboyCommunity/src/modules/login/styled.ts
@@ -1,0 +1,32 @@
+import { Pressable } from "react-native";
+import styled from "styled-components";
+
+export const ImportButton = styled(Pressable)`
+  padding: 8px 24px;
+  border-color: black;
+  color: white;
+  border-radius: 8px;
+`;
+
+export const CreateAccountButton = styled(Pressable)`
+  border-radius: 8px;
+  padding: 8px 24px;
+  border-color: black;
+  border: 1px;
+  color: white;
+`;
+
+export const LoginButton = styled(Pressable)`
+  border-radius: 8px;
+  border: 1px;
+  background-color: gray;
+  padding: 8px 24px;
+  color: white;
+`;
+
+export const SkipButton = styled(Pressable)`
+  border-radius: 8px;
+  padding: 8px 24px;
+  border-color: black;
+  color: white;
+`;
