+++ JoyboyCommunity/src/modules/notifications/Notification/styled.ts
@@ -1,0 +1,8 @@
+import { View } from "react-native";
+import styled from "styled-components";
+
+export const NotificationLayout = styled(View)`
+  flex-direction: row;
+  gap: 18px;
+  padding: 0px 12px;
+`;
