+++ JoyboyCommunity/src/modules/notifications/index.tsx
@@ -1,32 +1,30 @@
-import { View, Text, Platform } from "react-native";
+import { View } from "react-native";
 import React from "react";
 import LayoutContainer from "../../components/skeleton/LayoutContainer";
 import ScrollableContainer from "../../components/skeleton/ScrollableContainer";
 import { testPostData } from "../../shared/data/testData";
 import Divider from "../../components/divider/Divider";
-import Post from "../../shared/components/Post";
-import Notification from "./components/Notification";
+import { Notification } from "./Notification";
 
-export default function Notifications() {
+export const Notifications: React.FC = () => {
   return (
     <LayoutContainer title="Notifications">
       <ScrollableContainer
         contentContainerStyle={{
           // width: "100%"
-          width: Platform.OS != "android" ? "100%" : 250,
-
+          width: "100%",
           flex: 1,
         }}
       >
         <View style={{ paddingVertical: 12, flex: 1, gap: 18, width: "100%" }}>
-          {testPostData.map((post,i) => (
+          {testPostData.map((post, i) => (
             <View key={post.id} style={{ gap: 12 }}>
               <Notification
                 post={{
                   author: post?.pubkey,
-                  content:post?.content,
-                  username:post?.pubkey,
-                  id:post?.id,
+                  content: post?.content,
+                  username: post?.pubkey,
+                  id: post?.id,
                 }}
               />
               <Divider />
@@ -36,4 +34,4 @@ export default function Notifications() {
       </ScrollableContainer>
     </LayoutContainer>
   );
-}
+};
