+++ JoyboyCommunity/src/modules/profile/index.tsx
@@ -1,24 +1,22 @@
 import { View, Image, StyleSheet, Platform, Pressable } from "react-native";
-import React, { useEffect, useState } from "react";
-import styled, { useTheme } from "styled-components/native";
+import React, { useEffect, useMemo, useState } from "react";
+import { useTheme } from "styled-components/native";
+import { Event as EventNostr } from "nostr-tools";
+import { TouchableOpacity } from "react-native-gesture-handler";
 import Typography from "../../components/typography";
-import {
-  FlatList,
-  ScrollView,
-  TouchableOpacity,
-} from "react-native-gesture-handler";
-import Divider from "../../components/divider/Divider";
-import { testPostData } from "../../shared/data/testData";
-import Post from "../../shared/components/Post";
 import { useWindowDimensions } from "react-native";
 import { TabView, SceneMap } from "react-native-tab-view";
-import ScrollableContainer from "../../components/skeleton/ScrollableContainer";
-import { useBottomTabBarHeight } from "@react-navigation/bottom-tabs";
 import { useLocalstorage } from "../../hooks/useLocalstorage";
 import { useNostr } from "../../hooks/useNostr";
-import { Event as EventNostr } from "nostr-tools";
 import { INoteRepostParsed, IUserEvent } from "../../types";
 import { filterRepliesOnEvents } from "../../utils/filter";
+import {
+  NotesRoute,
+  RepliesRoute,
+  RepostsRoute,
+  ReactionsRoute,
+} from "./routes";
+import styles from "./styles";
 
 export default function Profile() {
   const theme = useTheme();
@@ -47,114 +45,6 @@ export default function Profile() {
     { key: "reposts", title: "Reposts" },
   ]);
 
-  const NotesRoute = () => {
-    return (
-      <FlatList
-        contentContainerStyle={{
-          paddingTop: 16,
-          // paddingBottom: bottomBarHeight,
-        }}
-        data={events}
-        keyExtractor={(item) => item?.id}
-        renderItem={({ item }) => {
-          return (
-            <Post
-              //  post={item}
-              sourceUser={profile?.picture}
-              event={item}
-            />
-          );
-        }}
-        ItemSeparatorComponent={() => (
-          <View style={{ marginVertical: 18 }}>
-            <Divider />
-          </View>
-        )}
-      />
-    );
-  };
-  /** TODO fix issue multi renders replies */
-  const RepliesRoute = () => {
-    return (
-      <FlatList
-        contentContainerStyle={{
-          paddingTop: 16,
-        }}
-        data={replies}
-        // keyExtractor={(item) => item?.id}
-        keyExtractor={(item, index) => index?.toString()}
-
-        renderItem={({ item }) => {
-          return <Post event={item}
-            sourceUser={profile?.picture}
-          />;
-        }}
-        ItemSeparatorComponent={() => (
-          <View style={{ marginVertical: 18 }}>
-            <Divider />
-          </View>
-        )}
-      />
-    );
-  };
-
-  const RepostRoute = () => {
-    // const bottomBarHeight = useBottomTabBarHeight();
-
-    return (
-      <FlatList
-        contentContainerStyle={{
-          paddingTop: 16,
-          // paddingBottom: bottomBarHeight,
-        }}
-        data={reposts}
-        keyExtractor={(item) => item?.event?.id}
-        renderItem={({ item }) => {
-          return (
-            <Post
-              // post={item}
-              sourceUser={profile?.picture}
-              event={item?.event}
-              repostedEvent={item?.repost}
-            />
-          );
-        }}
-        ItemSeparatorComponent={() => (
-          <View style={{ marginVertical: 18 }}>
-            <Divider />
-          </View>
-        )}
-      />
-    );
-  };
-
-  const ReactionsRoute = () => {
-    // const bottomBarHeight = useBottomTabBarHeight();
-    return (
-      <FlatList
-        contentContainerStyle={{
-          paddingTop: 16,
-          // paddingBottom: bottomBarHeight,
-        }}
-        data={reactions}
-        keyExtractor={(item) => item?.id}
-        renderItem={({ item }) => {
-          return (
-            <Post
-              //  post={item}
-              sourceUser={profile?.picture}
-              event={item}
-            />
-          );
-        }}
-        ItemSeparatorComponent={() => (
-          <View style={{ marginVertical: 18 }}>
-            <Divider />
-          </View>
-        )}
-      />
-    );
-  };
   // Fetch user based on userId pubkey
   useEffect(() => {
     const isConnectedUser = async () => {
@@ -168,7 +58,7 @@ export default function Profile() {
           setIsConnected(true);
           setPublicKey(publicKeyConnected);
         }
-      } catch (e) { }
+      } catch (e) {}
     };
 
     isConnectedUser();
@@ -182,7 +72,7 @@ export default function Profile() {
         if (publicKey && !isLoading && !profile && !isFirstLoadDone) {
           await handleMyProfileByPublicKey();
         }
-      } catch (e) { }
+      } catch (e) {}
     };
     handleInfo();
   }, [publicKey, isLoading, profile, isFirstLoadDone]);
@@ -211,13 +101,12 @@ export default function Profile() {
           let contentParsed = JSON.parse(userQueryReq?.content);
           let profile: IUserEvent = contentParsed;
           setProfile(profile);
-        } catch (e) { }
+        } catch (e) {}
 
         let events = await getEventsNotesFromPubkey(publicKey, [
           1, // note
           6, // repost
           7, // reactions
-
         ]);
 
         let notesAllTags = events?.filter((e) => e?.kind == 1);
@@ -241,13 +130,13 @@ export default function Profile() {
         });
         let reactions = events?.filter((e) => e?.kind == 7);
         // let replies = filterRepliesOnEvents(events)
-        let repliesFilter = filterRepliesOnEvents(notesAllTags)
+        let repliesFilter = filterRepliesOnEvents(notesAllTags);
 
         setReplies(repliesFilter);
         setReactions(reactions);
         setReposts(reposts);
         console.log("replies", repliesFilter);
-        let notes = notesAllTags?.filter((n) => n?.tags?.length == 0 )
+        let notes = notesAllTags?.filter((n) => n?.tags?.length == 0);
         // let notes = notesAllTags?.filter((n) => n?.tags?.length == 0 || !n?.tags?.find(e => e?.includes("e")))
         console.log("notes", notes);
         // console.log("reposts", reposts);
@@ -300,12 +189,17 @@ export default function Profile() {
       </View>
     );
   };
-  const renderScene = SceneMap({
-    posts: NotesRoute,
-    reactions: ReactionsRoute,
-    reposts: RepostRoute,
-    replies: RepliesRoute,
-  });
+
+  const renderScene = useMemo(() => {
+    return SceneMap({
+      posts: () => <NotesRoute events={events} profile={profile} />,
+      reactions: () => (
+        <ReactionsRoute reactions={reactions} profile={profile} />
+      ),
+      reposts: () => <RepostsRoute reposts={reposts} profile={profile} />,
+      replies: () => <RepliesRoute replies={replies} profile={profile} />,
+    });
+  }, [profile, events, reactions, reposts, replies]);
 
   return (
     <View style={{ flex: 1 }}>
@@ -313,7 +207,7 @@ export default function Profile() {
         <Image
           source={{ uri: "https://picsum.photos/200/300" }}
           style={{
-            width: Platform.OS != "android" ? "100%" : 250,
+            width: "100%",
 
             height: 200,
             resizeMode: "cover",
@@ -371,23 +265,3 @@ export default function Profile() {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  tabBar: {
-    width: Platform.OS != "android" ? "100%" : 250,
-
-    paddingHorizontal: 4,
-    flexDirection: "row",
-    borderBottomColor: "#e4e4e7",
-    borderBottomWidth: 1,
-  },
-  tabItem: {
-    marginHorizontal: 8,
-    paddingVertical: 8,
-    paddingHorizontal: 4,
-    alignItems: "flex-start",
-  },
-});
